% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignCodexWithScRNAseq.R
\name{alignScRNAseqToCodex}
\alias{alignScRNAseqToCodex}
\title{Align the single cell RNAseq data to codex data}
\usage{
alignScRNAseqToCodex(
  codexSeu,
  scSeu,
  classifier,
  newAssayName = "aligned",
  assay = "RNA"
)
}
\arguments{
\item{codexSeu}{The Codex Seurat object}

\item{scSeu}{The single cell RAN-seq Seurat object}

\item{classifier}{A KNN classifer created from the scRNA-seq Seurat object by
\link{buildClassifierFromSeurat}}

\item{assay, newAssayName}{The assay name for single cell RNAseq expressions.}
}
\value{
An aligned Seurat object.
}
\description{
Align the scRNA-seq Seurat object to the Codex data by the classifer
trained from the scRNA-seq Seurat object by \link{buildClassifierFromSeurat}
}
\examples{
\dontrun{
markers <- readRDS(system.file('extdata', 'markers.name.map.rds',
package='cdQuPath'))
pseudo <- readRDS(system.file('extdata', 'pseudobulk.scRNAseq.rds',
package='cdQuPath'))
classifier <- buildClassifierFromSeurat(pseudo, markers)
tsv <- system.file("extdata", "test.qptiff.tsv.zip",
                   package = "cdQuPath",
                   mustWork = TRUE)
seu <- createSeuratObj(tsv, useValue = 'Median',
                       markerLocations = CodexPredefined$markerLocations)
seu <- fitGMM(seu)
alignedPseudo <- alignScRNAseqToCodex(codexSeu = seu,
                    scSeu = pseudo,
                    classifier = classifier)
}
}
