% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknn.R
\name{buildClassifier}
\alias{buildClassifier}
\title{Build a generalized k-Nearest Neighbors Classifier for given reference}
\usage{
buildClassifier(data, features, celltypes)
}
\arguments{
\item{data}{A data.frame, containing features for \link[e1071]{gknn}
training. Data must be normalized into the range of 0 to 1.}

\item{features}{The features used for cell type prediction.}

\item{celltypes}{The cell types for each row or column of the data.}
}
\value{
An \link[e1071]{gknn} model.
}
\description{
For a given features data, comutes the k-nearest neighbours classifier
making use of general distance.
}
\examples{
data(iris)
iris[, 1:4] <- apply(iris[, 1:4], 2, function(.ele){
  (.ele-min(.ele))/max(.ele)
})
classifier <- buildClassifier(iris, colnames(iris)[1:4], iris$Species)
predictCelltypes(classifier, iris[c(1, 51, 101), ])
}
