% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignCodexWithScRNAseq.R
\name{buildClassifierFromSeurat}
\alias{buildClassifierFromSeurat}
\title{create a KNN classifier by a given Seurat object}
\usage{
buildClassifierFromSeurat(
  seuObj,
  markerList,
  doLog2transform = TRUE,
  assay = "RNA",
  layer = "count"
)
}
\arguments{
\item{seuObj}{An Seurat object}

\item{markerList}{A list of protein-marker map}

\item{doLog2transform}{logical(1L). Do log2(x+1) transform or not for the
'RNA' assay.}

\item{assay, layer}{The assay and layer used to create the classifier.}
}
\value{
An \link[e1071]{gknn} object
}
\description{
Prepare the KNN classifier from a given Seurat object.
}
\examples{
\dontrun{
markers <- readRDS(system.file('extdata', 'markers.name.map.rds',
package='cdQuPath'))
pseudo <- readRDS(system.file('extdata', 'pseudobulk.scRNAseq.rds',
package='cdQuPath'))
library(future.apply)
plan(multisession)
buildClassifierFromSeurat(pseudo, markers)
}
}
