% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{CodexPredefined}
\alias{CodexPredefined}
\title{Predefined variables}
\format{
An object of class \code{list} of length 9.
}
\usage{
CodexPredefined
}
\description{
Predefined variables
}
\examples{
CodexPredefined$markerLocations

}
\keyword{datasets}
