% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coOccurence.R
\name{coOccurrence}
\alias{coOccurrence}
\title{Get co-occurrence for given bins}
\usage{
coOccurrence(
  seu,
  anno_col = "celltype",
  minRadius = 0,
  maxRadius = 1000,
  bins = 100,
  spatialCoordsNames = CodexPredefined$spatialCoordsNames,
  celldist,
  ...
)
}
\arguments{
\item{seu}{An seurat object}

\item{anno_col}{The annotation column used to define the factors.}

\item{minRadius, maxRadius}{Minimal and maximal radius}

\item{bins}{The number of the bins}

\item{spatialCoordsNames}{The column names of coordinates}

\item{celldist}{Output of \link{findNearCellsByRadius}}

\item{...}{not used.}
}
\value{
A data matrix contain the cell proportion and co-occurrence for
each bin.
}
\description{
Calculate co-occurrence for each bin given range.
}
\examples{
# example code

}
