% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSeuratObj.R
\name{createSeuratObj}
\alias{createSeuratObj}
\title{Create Seurat object}
\usage{
createSeuratObj(
  path,
  useValue = c("Mean", "Median", "Min", "Max", "Std.Dev"),
  markerLocations = "Cell",
  ...
)
}
\arguments{
\item{path}{character(1L) or connection. The path to measurements exported by
ExportCellDetectionMeasurement.groovy in QuPath.}

\item{useValue}{character(1L). The values should be used as count table.}

\item{markerLocations}{The marker locations for each marker or charcter(1L)
for all markers. The possible locations are 'Nucleus', and 'Cell'.
'Cytoplasm' and 'Membrane' are also available but not suggested.}

\item{...}{parameters could be used by \link[SeuratObject]{CreateSeuratObject}}
}
\value{
An Seurat object.
}
\description{
Create Seurat object from output of ExportCellDetectionMeasurement.groovy
}
\examples{
# example code
tsv <- system.file("extdata", "test.qptiff.tsv.zip",
                   package = "cdQuPath",
                   mustWork = TRUE)
seu <- createSeuratObj(tsv, useValue = 'Median',
                       markerLocations = CodexPredefined$markerLocations)
seu
}
