% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCellType.R
\name{detectCellType}
\alias{detectCellType}
\title{Assign cell type}
\usage{
detectCellType(
  seu,
  classifier = CodexPredefined$classifier,
  method = c("Rank", "Boolean"),
  cutoff = 0.5,
  ...
)
}
\arguments{
\item{seu}{A Seurat object.}

\item{classifier}{A character vector of the cell type assignment markers
or a list of celltypes.
If it is a character vector,
the names of the classifier is the cell type and the content of the
classifier is the markers. The order of the classifier is the priority of
assignment for \code{celltype_first} column in output metadata.
The list example please see CodexPredefined$classifier2}

\item{method}{The cell type assignment method.
First the function will convert the probabilities
to positive or negative value by cutoff 0.5.
Option 'Rank' will use the rank of probabilities for the positive markers.
And option 'Boolean' will assign the cell type
by the order of classifier for the positive markers.}

\item{cutoff}{The cutoff value for probabilities.}

\item{...}{Not used.}
}
\value{
A Seurat object with metadata
\code{celltype} (top celltype),
\code{celltype_all} (all celltype),
\code{celltype_first} (the first assignment by the pritority of the classifier).
}
\description{
Assign cell types by given cell markers.
}
