% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamicRanges.R
\name{dynamicRanges}
\alias{dynamicRanges}
\title{Calculate the dynamic range.}
\usage{
dynamicRanges(seu, topN = 20, bottomN = 0.1, layer = "counts", ...)
}
\arguments{
\item{seu}{A Seurat object.}

\item{topN, bottomN}{A numeric.
If it is greater than 1, use top topN brightest cells or bottomN weakest
cells.
If it is smaller than 1 and greater than 0, use top 100*topN/bottomN
percentage cells.}

\item{layer}{The matrix used to calculate the dynamic range.}

\item{...}{parameters for GetAssayData}
}
\value{
A list with element brightest, weakest, and dynamic_range.
For dynamic ranges, Inf indicates zero background. NaN indicates zero/zero
}
\description{
The standard approach assessing dynamic range is to calculate a
signal-to-background (SNR) ratio by dividing the average of the top 20
brightest cells by the average intensity of the weakest 10\% of cells.
An SNR of 10 or more supports reliable image analysis.
The recommend SNR range is (10, Inf], typically > 100.
Lower than 3 indicates poor-performing antibodies.
}
\examples{
# example code

}
