% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearCellsByRadius.R
\name{findNearCellsByRadius}
\alias{findNearCellsByRadius}
\title{Find the nearby cells by a given distance}
\usage{
findNearCellsByRadius(
  seu,
  maxRadius,
  spatialCoordsNames = CodexPredefined$spatialCoordsNames
)
}
\arguments{
\item{seu}{An seurat object.}

\item{maxRadius}{numeric(1L). Maximal Euclidean distance.}

\item{spatialCoordsNames}{The column names of coordinates}
}
\value{
A data.frame of distance pairs with column name
as 'cell_1', 'cell_2', 'dist'
}
\description{
Using RANN::nn2 function to get the cell-cell distance by a given distance.
}
\examples{
# example code

}
