% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratUtilities.R
\name{findVarMarkers}
\alias{findVarMarkers}
\title{Find variable markers}
\usage{
findVarMarkers(seu, avoidMarkers, nFeatures = 10L, ...)
}
\arguments{
\item{seu}{A Seurat object}

\item{avoidMarkers}{The antibody markers should be avoid.
eg avoidMarkers = c('sox9', 'hist3h2a').}

\item{nFeatures}{Number of features to select as top variable features;
only used when selection.method is set to 'dispersion' or 'vst'. See
\link[Seurat]{FindVariableFeatures}.}

\item{...}{other parameters used by \link[Seurat]{FindVariableFeatures}.}
}
\value{
A Seurat object
}
\description{
Identifies markers that are used for cluster analysis by
\link[Seurat]{FindVariableFeatures}
}
\examples{
# example code

}
