% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratUtilities.R
\name{fitGMM}
\alias{fitGMM}
\title{Rescale data using GMM}
\usage{
fitGMM(seu, ...)
}
\arguments{
\item{seu}{A Seurat object}

\item{...}{Not use.}
}
\value{
A Seurat object with a new assay named as 'GMM'. The re-scaled values
saved in layer 'data'.
}
\description{
Re-scale data using multiclass Gaussian Mixture Model. If data is lack of
variance, the function will trying to do log normal distribution re-scale.
Re-scaled values will be probabilities between 0-1.
}
