% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTOM.R
\name{getTOM}
\alias{getTOM}
\title{Topological overlap matrix similarity}
\usage{
getTOM(
  seu,
  softPower = 2,
  dofilter = TRUE,
  assay = CodexPredefined$defaultAssay,
  output = c("plot", "matrix"),
  hclust_method = "complete",
  ...
)
}
\arguments{
\item{seu}{A Seurat object}

\item{softPower}{Soft thresholding power.}

\item{dofilter}{logical(1L). Filter the expression data or not.}

\item{assay}{The expression assay name. Do NOT use GMM.}

\item{output}{The output of the function. Available options: plot (default),
and matrix.}

\item{hclust_method}{The agglomeration method to be used.
See \link[stats]{hclust}.}

\item{...}{Parameters for \link[WGCNA]{adjacency}.}
}
\value{
A ggplot object or a dist matrix.
}
\description{
Calculate TOM by WGCNA
}
