% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ggHeatmap}
\alias{ggHeatmap}
\title{plot heatmap by ggplot2}
\usage{
ggHeatmap(ht, title, hclust_method = "complete", ...)
}
\arguments{
\item{ht}{A data frame with three columns, indicates the x, y, and value}

\item{title}{The title}

\item{hclust_method}{The agglomeration method to be used.
See \link[stats]{hclust}.}

\item{...}{Parameter passed to \link[stats]{dist}}
}
\description{
plot heatmap by ggplot2
}
