% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R
\name{JaccardIndex}
\alias{JaccardIndex}
\title{Calculate Jaccard Index}
\usage{
JaccardIndex(
  seu,
  anno_col,
  maxRadius,
  spatialCoordsNames = c("Cell.X", "Cell.Y"),
  celldist,
  ...
)
}
\arguments{
\item{seu}{An object of Seurat.}

\item{anno_col}{The annotation column used to define the factors.}

\item{maxRadius}{Maximal Euclidean distance of the neighbors.}

\item{spatialCoordsNames}{The column names of coordinates}

\item{celldist}{Output of \link{findNearCellsByRadius}}

\item{...}{not used.}
}
\value{
A data.frame with column names of 'A', 'B' and 'JaccardIndex'
}
\description{
Calculate Jaccard Index
}
