% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratUtilities.R
\name{loessSmooth}
\alias{loessSmooth}
\title{Normalize the signal by Loess}
\usage{
loessSmooth(seu, markers, pseudoCounts = 1, ...)
}
\arguments{
\item{seu}{A Seurat object}

\item{markers}{The markers to be normalized.}

\item{pseudoCounts}{Numeric(1L) to avoid log2(0).}

\item{...}{Parameters passed to \link{normalize.loess}.}
}
\value{
A Seurat object with normalized values saved in layer 'counts'.
The raw counts saved as a new Assay named as raw_counts.
}
\description{
Normalize the sigal by locally polynormial regression fitting.
}
