% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoodsAna.R
\name{neighborhoodsAna}
\alias{neighborhoodsAna}
\title{Neighborhoods analysis}
\usage{
neighborhoodsAna(
  seu,
  k = 25,
  anno_col = "celltype_first",
  clusterK = 12,
  spatialCoordsNames = CodexPredefined$spatialCoordsNames,
  ...
)
}
\arguments{
\item{seu}{A Seurat object}

\item{k}{The maximum number of nearest cells to compute.
See \link[hoodscanR]{findNearCells}.}

\item{anno_col}{Character vector. The name of annotation column to use.
See \link[hoodscanR]{findNearCells}.}

\item{clusterK}{The number of clusters for \link[hoodscanR]{clustByHood}}

\item{spatialCoordsNames}{The column names for spatial coordinates in the
metadata of the Seurat object.}

\item{...}{Other parameters used by \link[hoodscanR]{clustByHood} except
'object', 'pm_cols' and 'k'.}
}
\description{
Use hoodscanR package to do the neighborhoods analysis.
}
