% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratUtilities.R
\name{normData}
\alias{normData}
\title{Do normalization}
\usage{
normData(
  seu,
  normalizationMethod = c("LogNormalize", "CLR", "RC", "zscore"),
  ...
)
}
\arguments{
\item{seu}{A Seurat object}

\item{normalizationMethod}{Normalization method for data table. Available
methods are 'LogNormalize', 'CLR' (centered log ratio), 'RC' (CPM), 'zscore'.}

\item{...}{Parameters can be used by \link[Seurat]{NormalizeData}}
}
\value{
A Seurat object
}
\description{
Do normalization for the Seurat object created by \link{createSeuratObj}
}
