% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R
\name{plotJaccardIndex}
\alias{plotJaccardIndex}
\title{Plot a heatmap for JaccardIndex output}
\usage{
plotJaccardIndex(
  ji,
  title = "neighborhood Jaccard Index",
  hclust_method = "complete",
  ...
)
}
\arguments{
\item{ji}{output of \link{JaccardIndex}}

\item{title}{The title.}

\item{hclust_method}{The agglomeration method to be used.
See \link[stats]{hclust}.}

\item{...}{Parameter passed to \link[stats]{dist}}
}
\value{
A ggplot object
}
\description{
Plot a heatmap for JaccardIndex output
}
