% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTOM.R
\name{plotModules}
\alias{plotModules}
\title{Plot Modules}
\usage{
plotModules(TOM)
}
\arguments{
\item{TOM}{Topological overlap matrix.}
}
\description{
Plot a hierarchical clustering dendrogram and the modules by a given TOM
}
