% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gknn.R
\name{predictCelltypes}
\alias{predictCelltypes}
\title{predict celltypes by gknn classifier}
\usage{
predictCelltypes(classifier, data)
}
\arguments{
\item{classifier}{a gknn classifier exported by \link{buildClassifier}.}

\item{data}{A data.frame, containing features for \link[e1071]{gknn}
prediction.}
}
\value{
Celltype factors for each cell.
}
\description{
classify putative cell types by given classifier.
}
\examples{
data(iris)
iris[, 1:4] <- apply(iris[, 1:4], 2, function(.ele){
  (.ele-min(.ele))/max(.ele)
})
classifier <- buildClassifier(iris, colnames(iris)[1:4], iris$Species)
predictCelltypes(classifier, iris[c(1, 51, 101), ])

}
