% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialScore.R
\name{spatialScore}
\alias{spatialScore}
\title{Calculate spatial score}
\usage{
spatialScore(spe, celltypeCenter, celltypeLeft, celltypeRight, ...)
}
\arguments{
\item{spe}{Output of \link{neighborhoodsAna}. An object of SpatialExperiment.}

\item{celltypeCenter, celltypeLeft, celltypeRight}{The celltype for center,
left and right.}

\item{...}{not used.}
}
\value{
A data.frame with spatialScore, distance to Right celltype and
distance to Left celltype.
}
\description{
Calculate spatial score
}
