% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndex.R
\name{JaccardIndex}
\alias{JaccardIndex}
\title{Calculate Jaccard Index}
\usage{
JaccardIndex(spe, maxgap = Inf, ...)
}
\arguments{
\item{spe}{Output of \link{neighborhoodsAna}. An object of SpatialExperiment.}

\item{maxgap}{Maximal distance of the neighbors.}

\item{...}{not used.}
}
\description{
Calculate Jaccard Index
}
