% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratUtilities.R
\name{markerCorrelation}
\alias{markerCorrelation}
\title{create marker correlation heatmap}
\usage{
markerCorrelation(
  seu,
  layers = Layers(seu),
  method = c("spearman", "pearson", "kendall"),
  hclust_method = "complete",
  output = c("plot", "matrix"),
  ...
)
}
\arguments{
\item{seu}{A Seurat object}

\item{layers}{Layers of the default assay.}

\item{method}{A character string indicating which correlation coefficient
is to be computed. See \link[stats]{cor}.}

\item{hclust_method}{The agglomeration method to be used.
See \link[stats]{hclust}.}

\item{output}{The output of the function. Available options: plot (default),
and matrix.}

\item{...}{Parameter passed to \link[stats]{dist}}
}
\value{
A list of ggplot object or a list of dist matrix.
}
\description{
Create marker correlation heatmaps clustered by hclust.
}
