% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialScore.R
\name{spatialScore}
\alias{spatialScore}
\title{Calculate spatial score}
\usage{
spatialScore(
  seu,
  celltypeCenter,
  celltypeLeft,
  celltypeRight,
  anno_col = "celltype",
  maxRadius = 1000,
  spatialCoordsNames = CodexPredefined$spatialCoordsNames,
  celldist,
  ...
)
}
\arguments{
\item{seu}{An seurat object}

\item{celltypeCenter, celltypeLeft, celltypeRight}{The celltype for center,
left and right.}

\item{anno_col}{The annotation column used to define the factors.}

\item{maxRadius}{Minimal and maximal radius}

\item{spatialCoordsNames}{The column names of coordinates}

\item{celldist}{Output of \link{findNearCellsByRadius}}

\item{...}{not used.}
}
\value{
A data.frame with spatialScore, distance to Right celltype and
distance to Left celltype.
}
\description{
Calculate spatial score
}
